requirements.txt
Brotli==1.0.9
click==8.1.2
cycler==0.11.0
dash==2.3.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.1.1
Flask-Compress==1.11
fonttools==4.33.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
kaleido==0.2.1
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
plotly==5.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
tenacity==8.0.1
Werkzeug==2.1.1
zipp==3.8.0
